% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{binder_builder}
\alias{binder_builder}
\title{Builds binder}
\usage{
binder_builder(path = ".", hub = "mybinder.org", urlpath = "rstudio")
}
\arguments{
\item{path}{path to local git controlled folder}

\item{hub}{mybinder.org (often used synonymously with Binder) is a single deployment of a BinderHub instance, managed by the creators of Binder. It serves as both a public service and a demonstration of the BinderHub technology but it is by no means the only BinderHub in existence. \href{https://binder.pangeo.io/}{Pangeo} (https://binder.pangeo.io/) is another instance of a BinderHub (pass "binder.pangeo.io" to the hub argument) that is hosted by the geoscience community and is free to use for anyone. It is quite likely that other BinderHubs will be available in the future. You can use the `hub` argument to set the BinderHub of your choice. If your organization wants to host one (with additional resources to serve your community), please read through the MyBinder documentation (https://mybinder.readthedocs.io )}

\item{urlpath}{Defaults to rstudio for a Rstudio server.}
}
\description{
This function kicks off the image build on Mybinder.org. Although it is not
entirely necessary to run this step, doing so will ensure that there is a
built image on binder ready to launch. Otherwise the first time (or if your
binder is infrequently used), the build can take a long time.
}
