% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_binder}
\alias{build_binder}
\title{Builds binder in the background (recommended over calling binder_builder directly)}
\usage{
build_binder(path = ".", hub = "mybinder.org", urlpath = "rstudio")
}
\arguments{
\item{path}{path to local Git controlled folder}

\item{hub}{The binder hub you wish to use. The default is mybinder but you can also try \href{https://binder.pangeo.io/}{Pangeo}}

\item{urlpath}{Defaults to rstudio for a Rstudio server.}
}
\description{
This function builds binder in the background and once an image is ready,
will open the Binder URL. You are feel to kill this process anytime and the build
still continue on the server.
}
