% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{generate_badge}
\alias{generate_badge}
\title{Generates a Binder badge to add to your README}
\usage{
generate_badge(path = ".", hub = "mybinder.org", urlpath = "rstudio")
}
\arguments{
\item{path}{Assumes local folder unless you say otherwise}

\item{hub}{The binder hub you wish to use. The default is mybinder but you can also try \href{https://binder.pangeo.io/}{Pangeo}}

\item{urlpath}{For R users we recommend the default Rstudio. If you wish to use vanilla Jupyter, leave this blank.}
}
\description{
Generates a Binder badge to add to your README
}
\examples{
# generate_badge("/path/to/repo")
}
