% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{last_modification_date}
\alias{last_modification_date}
\title{Last Modified late}
\usage{
last_modification_date(path = ".")
}
\arguments{
\item{path}{}
}
\description{
Returns the latest date for when a file was touched in a project
}
