% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_compendium_description.R
\name{write_compendium_description}
\alias{write_compendium_description}
\title{Creates a description file for a compendium}
\usage{
write_compendium_description(Type = "Compendium",
  Package = "Compendium title", Version = "0.0.1", path = ".")
}
\arguments{
\item{Type}{Default here is compendium}

\item{Package}{Name of your project}

\item{Version}{Version of your compendium}

\item{path}{path to project (in case it is elsewhere)}
}
\description{
Creates a description file for a compendium
}
