% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_installer.R
\name{write_install}
\alias{write_install}
\title{Write a install.R}
\usage{
write_install(path = ".")
}
\arguments{
\item{path}{Path to project}
}
\description{
This function writes a list of packages to install one per line in the form of `install.package('package_name')`
}
