% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_requirements.R
\name{write_requirements}
\alias{write_requirements}
\title{write_requirements}
\usage{
write_requirements(path = ".", requirements = NULL)
}
\arguments{
\item{path}{Path to project}

\item{requirements}{A list of Python packages that should be installed in your environment.}
}
\description{
Writes a list of Debian packages that should be installed before building
the Docker image
}
\examples{
write_requirements(path = ".",
requirements = "
numpy==1.16.*
matplotlib==3.*
seaborn==0.8.1
")
}
