% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_runtime.R
\name{write_runtime}
\alias{write_runtime}
\title{Write runtime.txt}
\usage{
write_runtime(dt = lubridate::ymd(lubridate::today()))
}
\arguments{
\item{dt}{Date you need R to lock down to}
}
\description{
Write runtime.txt
}
