% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_binder}
\alias{build_binder}
\title{Builds binder in the background (recommended over calling binder_builder directly)}
\usage{
build_binder(path = ".", hub = "mybinder.org", urlpath = "rstudio")
}
\arguments{
\item{path}{path to local Git controlled folder}

\item{hub}{mybinder.org (often used synonymously with Binder) is a single deployment of a BinderHub instance, managed by the creators of Binder. It serves as both a public service and a demonstration of the BinderHub technology but it is by no means the only BinderHub in existence. \href{https://binder.pangeo.io/}{Pangeo} (https://binder.pangeo.io/) is another instance of a BinderHub that is hosted by the geoscience community and is free to use for anyone. It is quite like that other BinderHubs will be available in the future. You can use the `hub` argument to set the BinderHub of your choice. If your organization wants to host one (with additional resources to serve your community), please read through the MyBinder documentation (https://mybinder.readthedocs.io ) \href{https://binder.pangeo.io/}{Pangeo}}

\item{urlpath}{Defaults to rstudio for a Rstudio server.}
}
\description{
This function builds binder in the background and once an image is ready,
will open the Binder URL. You are feel to kill this process anytime and the build
still continue on the server.
}
