% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{generate_badge}
\alias{generate_badge}
\title{Generates a Binder badge to add to your README}
\usage{
generate_badge(path = ".")
}
\arguments{
\item{path}{Assumes local folder unless you say otherwise}
}
\description{
Generates a Binder badge to add to your README
}
\examples{
# generate_badge("/path/to/repo")
}
