% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependencies.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Get project dependencies}
\usage{
get_dependencies(dir = ".")
}
\arguments{
\item{dir}{Root directory to search from}
}
\value{
vector
}
\description{
Get project dependencies
}
