% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dependencies.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Pulls out R dependencies from a folder.}
\usage{
get_dependencies(path = ".")
}
\arguments{
\item{path}{Path to search}
}
\description{
Pulls out R dependencies from a folder.
}
