% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_compendium_description.R
\name{write_compendium_description}
\alias{write_compendium_description}
\title{Creates a description file for a compendium}
\usage{
write_compendium_description(Type = "Compendium",
  Package = "Compendium title", Description = "Compendium Description",
  Version = "0.0.1", path = ".")
}
\arguments{
\item{Type}{Default here is compendium}

\item{Package}{Name of your compendium}

\item{Description}{Description of your compendium}

\item{Version}{Version of your compendium}

\item{path}{path to project (in case it is elsewhere)}
}
\description{
The idea behind a compendium is to have a minimal description file that makes
it easy for anyone to 'install' your analysis dependencies. This makes it
possible for someone to run your code easily.
}
