% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_compendium_description.R
\name{write_compendium_description}
\alias{write_compendium_description}
\title{Creates a description file for a compendium}
\usage{
write_compendium_description(type = "Compendium",
  package = "Compendium title", description = "Compendium description",
  version = "0.0.1", path = ".")
}
\arguments{
\item{type}{Default here is compendium}

\item{package}{Name of your compendium}

\item{description}{Description of your compendium}

\item{version}{Version of your compendium}

\item{path}{path to project (in case it is not in the current working directory)}
}
\description{
The idea behind a compendium is to have a minimal description file that makes
it easy for anyone to 'install' your analysis dependencies. This makes it
possible for someone to run your code easily.
}
\details{
To automatically populate author information, you may set usethis options in your `.rprofile` like so.
\code{options(
  usethis.full_name = "Karthik Ram",
  usethis.description = list(
  `Authors@R` = 'person("Karthik", "Ram", email = "karthik.ram@gmail.com", role = c("aut", "cre"),
  comment = c(ORCID = "0000-0002-0233-1757"))',
  License = "MIT + file LICENSE",
  Version = "0.0.0.9000"
  )
)}
}
