% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dockerfile.R
\name{write_dockerfile}
\alias{write_dockerfile}
\title{Write a minimal Dockerfile for a binder}
\usage{
write_dockerfile(base = "rocker/binder:latest", path = ".",
  method = "desc")
}
\arguments{
\item{base}{Rocker base}

\item{path}{path to binder repo}

\item{method}{If left blank, it will install using the DESCRIPTION file. Otherwise it will use the install.R file.}
}
\description{
Write a minimal Dockerfile for a binder
}
