% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dockerfile.R
\name{write_dockerfile}
\alias{write_dockerfile}
\title{Write a minimal Dockerfile for a binder}
\usage{
write_dockerfile(base = NULL, path = ".", maintainer = "karthik",
  r_version = "latest")
}
\arguments{
\item{base}{Rocker base}

\item{path}{path to binder repo}

\item{maintainer}{Maintainer of the Dockerfile}

\item{r_version}{Version of R you wish to use. Will default to latest for now.}
}
\description{
Write a minimal Dockerfile for a binder
}
