% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dockerfile.R
\name{write_dockerfile}
\alias{write_dockerfile}
\title{Write a minimal Dockerfile for a binder}
\usage{
write_dockerfile(base = NULL, path = ".", maintainer = "karthik",
  r_date = NULL)
}
\arguments{
\item{base}{Rocker base}

\item{path}{path to binder repo}

\item{maintainer}{Maintainer of the Dockerfile}

\item{r_date}{The date for which you'd like to lock down this project.
Projects that match current release date go to "latest"}
}
\description{
Write a minimal Dockerfile for a binder
}
