% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_runtime.R
\name{write_runtime}
\alias{write_runtime}
\title{Write runtime.txt}
\usage{
write_runtime(r_date = lubridate::ymd(lubridate::today()))
}
\arguments{
\item{r_date}{Date you need R to lock down to. By default it picks the most
recent date a file was touched in this project but you can override this by
specifying this explicitly.}
}
\description{
Write runtime.txt
}
